[REF] mail: split rtc_service
This commit split rtc_service class in multiple files.
