+++ addons/mail/static/tests/mail_test_helpers.js
@@ -75,7 +75,8 @@ import { ResRole } from "./mock_server/mock_models/res_role";
 import { ResUsers } from "./mock_server/mock_models/res_users";
 import { ResUsersSettings } from "./mock_server/mock_models/res_users_settings";
 import { ResUsersSettingsVolumes } from "./mock_server/mock_models/res_users_settings_volumes";
-import { Network, Rtc } from "@mail/discuss/call/common/rtc_service";
+import { Network } from "@mail/discuss/call/common/network";
+import { Rtc } from "@mail/discuss/call/common/rtc_model";
 import { UPDATE_EVENT } from "@mail/discuss/call/common/peer_to_peer";
 import { SoundEffects } from "@mail/core/common/sound_effects_service";
 import { DiscussAppCategory } from "@mail/discuss/core/public_web/discuss_app/discuss_app_category_model";
